cdc: learn about gossiped streams.
When a node learns that another node changes its state to NORMAL
or LEAVING, it will remember the CDC streams sent by that node
in a streams_metadata instance, similarily to how tokens
are remembered in a token_metadata instance.
We require that at least one stream is sent by such a node (this was
taken care of in previous commits in the stream generation
and gossiping code).

The streams are saved in streams_metadata before tokens are saved
in token_metadata. This is important so that when the node becomes
a coordinator of a write, it will know about at least one stream
of every node it will send the write to, hence it will be able
to send a corresponding CDC log mutation.

Similarly to tokens, the streams are saved in the PEERS table and
recovered when the node is restarting.