storage_service: wait for all seeds' tokens on the "fast join" path
When joining as a seed or with auto_bootstrap=off (NEVER DO THAT),
the node will wait until it learns (through the gossip protocol) about
the tokens of all other seeds, which it needs to know to propose a CDC
generation with a correct set of streams.

Before this commit, the node would:
1. wait for just one other seed's tokens,
2. wait for this seed to be in NORMAL status.

This commit both strenghtens and relaxes the above conditions:
1. now it waits for all seeds,
2. but it doesn't require them to be in NORMAL status. It is enough if
   they announce their tokens in other status as well. That's why the
   ANNOUNCING_TOKENS status was introduced in previous commits.

This enables boostrapping a new cluster in which all nodes are marked as
seeds by all nodes.

(but don't ever do that, please; not until safe concurrent cluster
bootstrap is actually implemented)