versioned_value: add ANNOUNCING_TOKENS status
This status will be used by a node which wishes to join a cluster
to tell other nodes what its chosen tokens will be, but without the
nodes inserting these tokens into their token rings (which would change
the sets of replicas calculated by these nodes when they become
coordinators) as is done with NORMAL, BOOT, or LEAVING statuses.

The node will use this status whenever it wants the other nodes
to somehow react to the fact that it's joining, before it begins
joining.