raft: fsm: fix assertion failure on stray rejects
When `raft::fsm` received an `append_reply::rejected` message as a
leader from a follower of the same term, it would assert that
the index of the next log entry to replicate to that follower
(`progress.next_idx`) is not equal to the last known match index of the
follower (`progress.match_idx`). However, this assumption would not
always be true, as the regression test included in this commit shows.

We fix the problem by strengthening `is_stray_reject`. The check that
was previously only made in `PIPELINE` case
(`rejected.non_matching_idx <= match_idx`) is now always performed and
we add a new check: `rejected.last_idx < match_idx`. We also strengthen
the assert.

The commit improves the documentation by explaining that
`is_stray_reject` may return false negatives.  We also precisely state
the preconditions and postconditions of `is_stray_reject`, give a more
formal definition of `progress.match_idx`, formally argue how the
postconditions of `is_stray_reject` follow from its preconditions
and Raft invariants, and argue why the (strengthened) assert
must always pass.