test: raft: randomized_nemesis_test: generalize `with_env_and_ticker`
Generalize the type of the callback: use a template parameter instead of
`noncopyable_function` and don't assume the return type of the callback.

This allows returning a result from `with_env_and_ticker`, e.g. for
performing analysis or logging the results after a part of the test that
used the environment and ticker have finished.